Splits SSLCertConfigLoader into it's own file
