+++ iocore/net/Makefile.am
@@ -137,6 +137,7 @@ libinknet_a_SOURCES = \
 	ProxyProtocol.h \
 	ProxyProtocol.cc \
 	Socks.cc \
+	SSLCertConfigLoader.cc \
 	SSLCertLookup.cc \
 	SSLClientUtils.cc \
 	SSLConfig.cc \
