+++ iocore/net/P_SSLClientUtils.h
@@ -21,17 +21,9 @@
 
 #pragma once
 
-#include "tscore/ink_config.h"
-#include "tscore/Diags.h"
-#include "P_SSLUtils.h"
-#include "P_SSLConfig.h"
+struct SSLConfigParams;
 
 #include <openssl/ssl.h>
-
-// BoringSSL does not have this include file
-#ifndef OPENSSL_IS_BORINGSSL
-#include <openssl/opensslconf.h>
-#endif
 
 // Create and initialize a SSL client context.
 SSL_CTX *SSLInitClientContext(const struct SSLConfigParams *param);
