+++ iocore/net/P_SSLConfig.h
@@ -41,6 +41,7 @@
 #include "YamlSNIConfig.h"
 
 #include "P_SSLUtils.h"
+#include "P_SSLCertLookup.h"
 
 struct SSLCertLookup;
 struct ssl_ticket_key_block;
