+++ iocore/net/SSLSessionCache.cc
@@ -25,6 +25,8 @@
 
 #include <cstring>
 
+#define SSL_MAX_SESSION_SIZE 256
+
 #define SSLSESSIONCACHE_STRINGIFY0(x) #x
 #define SSLSESSIONCACHE_STRINGIFY(x) SSLSESSIONCACHE_STRINGIFY0(x)
 #define SSLSESSIONCACHE_LINENO SSLSESSIONCACHE_STRINGIFY(__LINE__)
